# Avoid try catch for a 5x speedup