Node 6.11

Before:

```
lookup static route x 4,194,264 ops/sec ±0.83% (93 runs sampled)
lookup dynamic route x 690,927 ops/sec ±0.83% (91 runs sampled)
lookup long static route x 2,040,913 ops/sec ±0.90% (90 runs sampled)
find static route x 5,143,679 ops/sec ±0.89% (88 runs sampled)
find dynamic route x 705,003 ops/sec ±0.93% (88 runs sampled)
find long static route x 2,425,694 ops/sec ±1.02% (89 runs sampled)
```

After:

```
lookup static route x 20,374,370 ops/sec ±1.45% (89 runs sampled)
lookup dynamic route x 701,705 ops/sec ±1.01% (90 runs sampled)
lookup long static route x 4,815,807 ops/sec ±2.82% (91 runs sampled)
find static route x 28,975,395 ops/sec ±1.01% (90 runs sampled)
find dynamic route x 723,708 ops/sec ±0.87% (91 runs sampled)
find long static route x 7,572,324 ops/sec ±5.56% (87 runs sampled)
```