@mcollina 2017-07-04
On Node.js 8.1 is the reverse:

Before:
```
lookup static route x 20,874,688 ops/sec ±0.99% (89 runs sampled)
lookup dynamic route x 1,092,509 ops/sec ±1.08% (87 runs sampled)
lookup long static route x 4,392,481 ops/sec ±0.90% (92 runs sampled)
find static route x 26,441,581 ops/sec ±0.63% (91 runs sampled)
find dynamic route x 1,180,374 ops/sec ±0.84% (90 runs sampled)
find long static route x 6,553,732 ops/sec ±0.85% (90 runs sampled)
```

after:
```
lookup static route x 17,900,136 ops/sec ±0.94% (92 runs sampled)
lookup dynamic route x 550,454 ops/sec ±2.04% (87 runs sampled)
lookup long static route x 2,972,802 ops/sec ±10.68% (85 runs sampled)
find static route x 24,610,052 ops/sec ±1.93% (85 runs sampled)
find dynamic route x 566,117 ops/sec ±1.93% (89 runs sampled)
find long static route x 4,453,069 ops/sec ±14.43% (83 runs sampled)
```