@mcollina 2017-07-04
In node 8.2 (or whenever we ship full Turbofan in core) the problem goes away, because it will be all turbofan then. The degradation in Node 8 happen because the `fastDecode`  method is compiled by turbofan, whine the rest by crankshaft, and there is no inlining between the two optimizers.